+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added: Browser utilities `CConnoisseur.ascend` and `CConnoisseur.reincarnate`.
 
 ## 0.2.0 - 2021-10-01
 - Added: Explicit support for Firefox.